refactor(administration of products): add controller a request into directory
